pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "hfDeepResearch"
version = "0.1.0"
description = "A project for deep research using smolagents"
authors = [
    { name = "User" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.37.1",
    "beautifulsoup4>=4.12.3",
    "datasets>=2.21.0",
    "google-search-results>=2.4.2",
    "huggingface_hub>=0.23.4",
    "mammoth>=1.8.0",
    "markdownify>=0.13.1",
    "numexpr>=2.10.1",
    "numpy>=2.1.2",
    "openai>=1.52.2",
    "openpyxl",
    "pandas>=2.2.3",
    "pathvalidate>=3.2.1",
    "Pillow>=11.0.0",
    "puremagic>=1.28",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "python_pptx>=1.0.2",
    "Requests>=2.32.3",
    "tqdm>=4.66.4",
    "transformers>=4.46.0",
    "youtube_transcript_api>=0.6.2",
    "chess",
    "sympy",
    "pubchempy",
    "Bio",
    "scikit-learn",
    "scipy",
    "pydub",
    "PyPDF2",
    "python-pptx",
    "xlrd",
    "SpeechRecognition",
    "pdfminer-six>=20240706",
    "smolagents @ git+https://github.com/huggingface/smolagents.git",
]

[project.optional-dependencies]
dev = [
    "pytest"
]
